Refactor cleanse_csv_data
Issue #81

Refactor cleansing of cvs data. Add a function to cleanse a single row
(or return None if the row isn’t useful), then run this over all rows
and filter out any we’re not interested in.

Remove the functions that operate over the entire set of rows as
they’re no longer needed.