ui: add mapping of meeting target ids to names
The API returns us meeting data separately to names of meeting targets
(albeit in the same API response via the ?include=... mechanism).  So we
need to do a merge of the two data sets where we count meetings grouped
by target id, and then map target ids to target names, so that we can
give the bubbles text labels corresponding to actual people or
departments or organisations.

This is the first step which works with USE_API = false.  The next step
will parse the "included" section of the API response so that it works
with data directly from the API.