ui: add routes to the app
We want to be able to navigate between different visualisation views
for organisations, people, government offices, ministers etc. whilst
keeping the same page layout with the header at the top.  So set up
some React nested routes and use this.props.children to render the
right components at the right levels in the component hierarchy.