add .hound.yml
Previously we didn't have the right linters enabled - in particular,
HoundCI's ESLint and Python linters are in beta, so they were disabled
by default, and jshint doesn't seem to cope well with ES6, so we want
to explicitly disable jshint and enable eslint and python instead.
But for clarity, we explicitly enable or disable all linters.