ui: catch occasional errors when store is empty
This seems to be some strange race condition where state.data is not
always initialized by the time mapStateToProps is invoked, causing an
exception when attempting to call getIn() on it.  Maybe it only
happens when auto-reloading the app whilst a git rebase is running,
but I'm not really sure.  Either way, we should watch out for this by
logging a warning.